Refactored StudySettings into a separate file to be in line with the rest of the project.
