+++ tests/antares/services/api_services/test_study_api.py
@@ -27,7 +27,8 @@ from antares.exceptions.exceptions import (
 from antares.model.area import Area, AreaProperties, AreaUi
 from antares.model.binding_constraint import BindingConstraint, BindingConstraintProperties
 from antares.model.link import Link, LinkProperties, LinkUi
-from antares.model.settings import GeneralProperties, StudySettings
+from antares.model.settings.general import GeneralProperties
+from antares.model.settings.study_settings import StudySettings
 from antares.model.study import Study, create_study_api
 from antares.service.service_factory import ServiceFactory
